requirements.txt
certifi==2018.10.15
chardet==3.0.4
defusedxml==0.5.0
Django==2.1
django-allauth==0.38.0
django-ckeditor==5.6.1
django-js-asset==1.1.0
django-widget-tweaks==1.4.3
idna==2.7
oauthlib==2.1.0
Pillow==5.3.0
python3-openid==3.1.0
pytz==2018.7
requests==2.20.1
requests-oauthlib==1.0.0
slugify==0.0.1
Unidecode==1.0.23
urllib3==1.24.1
